Fjerner unødvendig interface - Legger på konsistenssjekk ved lagring av grunnlag
