+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/iay/tjeneste/migrering/MigreringRestTjeneste.java
@@ -28,8 +28,6 @@ import javax.ws.rs.core.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import no.nav.foreldrepenger.abakus.domene.iay.GrunnlagReferanse;
